requirements.txt
annotated-types==0.5.0
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pydantic==2.3.0
pydantic_core==2.6.3
requests==2.31.0
typing_extensions==4.6.1
urllib3==2.0.2
Django==4.2.5
djangorestframework==3.14.0
django-filter==23.2
django-cors-headers